Fix header file usage -- there were many unnecessary header files included in busybox.h which slowed compiles.  I left only what was needed and then fixed up all the apps to include their own header files.  I also fixed naming for pwd.h and grp.h functions.  Tested to compile and run with libc5, glibc, and uClibc.  -Erik

+++ tr.c
@@ -191,9 +191,9 @@ extern int tr_main(int argc, char **argv)
 			map(input, input_length, output, output_length);
 		}
 		for (i = 0; i < input_length; i++)
-			invec[input[i]] = TRUE;
+			invec[(int)input[i]] = TRUE;
 		for (i = 0; i < output_length; i++)
-			outvec[output[i]] = TRUE;
+			outvec[(int)output[i]] = TRUE;
 	}
 	convert();
 	return (0);
